Rename palette as monokai, will be replaced with serialized value in the future
