+++ editor/moxie/src/ui.rs
@@ -25,7 +25,7 @@ use moxie_ui::widgets::dock::{
     DockAreaStyle, DockLeaf, DockNode, DockTree,
     DockWindowDescriptor, Edge, WindowRegistry, dock,
 };
-use moxie_ui::{MoxieUiPlugin, palette};
+use moxie_ui::{MoxieUiPlugin, monokai};
 
 /// Wires feathers theming, the editor UI tree, and the per-frame
 /// timeline/playback/preview systems.
@@ -93,7 +93,7 @@ fn setup_editor_ui(
             Camera2d
             Camera {
                 order: 10,
-                clear_color: { palette::BASE[0] },
+                clear_color: { monokai::BASE[0] },
             }
             TrackViewportCamera
         ])
