+++ editor/moxie_ui/src/lib.rs
@@ -11,8 +11,8 @@ pub mod fold;
 pub mod icons;
 pub mod inspector;
 pub mod layout;
+pub mod monokai;
 pub mod motion;
-pub mod palette;
 pub mod reactive;
 pub mod theme;
 pub mod widgets;
